Modified the Model to fit the new requirements.
The Model now uses additional stacks for easy and convenient data
management.
The printHistory() method takes care of converting the data stored in
the historyStack from postfix notation, to infix notation.
This is required since the new requirements state that while the input
remains in postfix notation, the output needs to be in infix notation.