Implemented the remaining binary operations as well as Clear and Undo.
The Undo operations performs according to the requirements, where it
will delete the last digit of the input if the user was typing an input,
and remove the previous operation if the user was not typing.

The Clear operation essentially does the same thing the constructor
does.

The binary arithmetic operations are straightforward, except for the
divide operation which currently outputs an error message and invokes
the Clear method if division by 0 occurs.