Finished the implementation of the Model as per specifications.
Implemented all the operations (buttons) that are specified by the
requirements.
The factorial operation only works for whole numbers, and gives an error
otherwise.
The pi operation adds the constant pi to the stack as it is pressed
instead of adding it to the total input. An input such as 3pi is still
possible by pressing the number desired, followed by pi.