Added a way to check operator precedence and minimize number of brackets
Added a method that checks the precedence of the current operator.
The printHistroy method uses the precedence to decide how to add the
brackets to the history expression, minimizing the number of brackets in
the final expression.

Also fixed a minor error, where the Divide method was actually
multiplying instead of dividing :P