Fixed a bug similar to the previous commit.
The bug would occur when pressing any unary or binary operation as soon
as a new calculation starts. Fixed by setting the reset flag in the
enoughOperands methods.

Also changed the way the result is being checked whether or not its a
whole number. This now allows for very large or very small number to be
represented in scientific notation.