Made changes according to the program operation evaluation
1) The history now always displays an equal sign after an undo, for both
evaluated expressions and input.

2) When using the undo operation while the user types an input, after
the last digit has been removed, a 0 is shown in the input field instead
of a blank value.

3) The precedence of operators should now now be handled properly and
the brackets should be added correctly.

4) "Sample Test" button now has a clear() operation at the beginning of
the sequence, so it will always display just the one sample expression
when pressed.

5) The Factorial operation now handles calculations of large numbers
properly, and no longer crashes.

6) When using the undo operation while the user types an input, and the
input was negative, after the last digit is removed the input field
displays 0 instead of "-" (such as in the case of undoing "-5").

7) When a math error occurs, the user is now able to undo the last
operation to fix the error. Every other button press, while the error is
still being handled, will reset the calculator.