Added the infrastructure needed to be able to draw graphs.
The MathValue class holds the data for each value entered to the
calculator.
Each value is treated as a double[];
If the value is a constant the array has the same value at each point
(on graph, line parallel to y = 0).
If the value is a variable x, or any operation resulting from operating
on x, the operation is calculated on each point of x, from -50 to 50
(currently thats the domain).