Added the favorites view
The favorites view holds up to 10 (can be changed) favorite equations.
Equations can be either shown on the graph or deleted.

Also deleted GraphView class since it had no use.

Introduced additional constructor to the ButtonAdapter class, which
allows to bind a "graph" or "delete" button to a specific favorite
entry.